Created a simple bot for Discord
Currently, the bot is using simple YAML-files to respond to simple messages.

Created a simple plugin system where we want more complex functionality
for both detecting if the bot should respond or not and how to respond.
This will allow a more modular approach, and will make sure that we
don't need to mess around too much in the main bot code in order to add
simple functionality for the bot.